refactor: rename component
- add href to link
- use pre-formatted title
- use aliased imports'
