+++ src/components/TeacherComponents/MyLibraryProgrammeCard/MyLibraryProgrammeCard.stories.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import { Meta, StoryObj } from "@storybook/react";
 import { oakDefaultTheme, OakThemeProvider } from "@oaknational/oak-components";
 
+import MyLibraryProgrammeCard from "./MyLibraryProgrammeCard";
+
 import {
   completeUnitLessons,
   incompleteUnitLessons,
-} from "../MyLibraryUnitCard/MyLibraryUnitCard.stories";
-
-import MyLibrarySubjectCard from "./MyLibrarySubjectCard";
+} from "@/fixtures/teachers/myLibrary";
 
 const sampleUnits = [
   {
@@ -42,17 +42,18 @@ const sampleUnits = [
     yearTitle: "Year 11",
     saveTime: new Date().toISOString(),
     href: "/teachers/programmes/english-secondary-ks4-aqa/units/poetry-analysis-romanticism/lessons",
-    lessons: completeUnitLessons.slice(0, 3), // Fewer lessons for variety
+    lessons: completeUnitLessons.slice(0, 3),
     onSave: () => console.log("Unsave unit 3"),
     isSaved: true,
   },
 ];
 
-const meta: Meta<typeof MyLibrarySubjectCard> = {
-  component: MyLibrarySubjectCard,
+const meta: Meta<typeof MyLibraryProgrammeCard> = {
+  component: MyLibraryProgrammeCard,
   tags: ["autodocs"],
   args: {
-    keyStage: "KS4",
+    programmeTitle: "English Secondary KS4 (AQA)",
+    programmeSlug: "english-secondary-ks4-aqa",
     subject: "english",
     savedUnits: sampleUnits,
   },
@@ -63,12 +64,12 @@ const meta: Meta<typeof MyLibrarySubjectCard> = {
 
 export default meta;
 
-type Story = StoryObj<typeof MyLibrarySubjectCard>;
+type Story = StoryObj<typeof MyLibraryProgrammeCard>;
 
 export const Default: Story = {
   render: (args) => (
     <OakThemeProvider theme={oakDefaultTheme}>
-      <MyLibrarySubjectCard {...args} />
+      <MyLibraryProgrammeCard {...args} />
     </OakThemeProvider>
   ),
 };
